pyproject.toml

[project]
dependencies = [
  "openai>=1.50.0",
  "chromadb>=1.3.0",
  "faiss-cpu>=1.9.0",
  "numpy>=2.0.0",
  "pypdf>=4.0.0",
  "python-dotenv>=1.0.0",
  "tiktoken>=0.7.0",
]
description = "A beginner-friendly RAG learning repository"
name = "rag-beginner-basics"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[project.optional-dependencies]
dev = [
  "jupyter>=1.1.0",
  "matplotlib>=3.9.0",
  "scikit-learn>=1.5.0",
  "ipywidgets>=8.0.0",
  "ruff>=0.3.0",
  "pytest>=8.0.0",
  "black>=24.0.0",
]

[project.scripts]
rag-basics = "rag_basics.main:main"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
